refactor(daemon): split remote clients
